phase 2:3 (next is cleanup)

+++ ax/gui/axBitmap.h
@@ -20,11 +20,13 @@
 //----------
 
 #ifdef AX_LINUX
-  #include "platform/linux/axBitmapLinux.h"
+  //#include "platform/linux/axBitmapLinux.h"
+  #include "gui/linux/axBitmapLinux.h"
 #endif
 
 #ifdef AX_WIN32
-  #include "platform/win32/axBitmapWin32.h"
+  //#include "platform/win32/axBitmapWin32.h"
+  #include "gui/win32/axBitmapWin32.h"
 #endif
 
 
@@ -58,7 +60,7 @@ class axBitmap : public axBitmapImpl
         // malloc ?
         mBuffer = new char[size];
 
-        if (aData) axMemcpy(mBuffer,(int)aData,size);
+        if (aData) axMemcpy(mBuffer,/*(int)*/aData,size);
         else axMemset(mBuffer,0,size);
         return mBuffer;
       }
